requirements.txt
asyncio
db-dtypes
fastapi
google-adk
google-auth
google-cloud-bigquery 
numpy
openpyxl
pandas
pandas-gbq
psycopg2-binary
pydantic
pyjwt
pwdlib
python-dotenv
requests
sqlmodel
uuid
uvicorn